refactor(web): extract web host runtime and streamline exports
Move the shared web host runtime into `internal/webhost` so the
repository keeps a direct ownership boundary around `runner.html`,
`game.js`, and related host assets.

Assemble web runtime bundles from the engine template, embedded host
assets, `ispx.wasm`, and `wasm_exec.js` instead of staging host files
through `$GOPATH/bin/ispx`.

Rename the reusable runtime export command to `exportwebruntime`,
remove the public `exporttemplateweb` concept, and keep the template
export step internal to the build scripts.

Also remove XBuilder-specific hooks from the shared host runtime so the
base web runtime stays product-neutral.
