+++ cmd/gox/main.go
@@ -2,6 +2,7 @@ package main
 
 import (
 	"embed"
+	"fmt"
 	"strings"
 
 	"github.com/goplus/spx/v2/cmd/gox/pkg/command"
@@ -50,7 +51,7 @@ func main() {
 	// Initialize the Args field if not already initialized
 	err := cmd.RunCmd(appName, appName, cmd.Version, projectFS, "template/project", "project")
 	if err != nil {
-		println("failed to run cmd:", err)
+		fmt.Println("failed to run cmd:", err)
 		return
 	}
 }
