+++ cmd/gox/template/platform/webminiprogram/index.html
@@ -5,7 +5,7 @@
 	<meta charset="utf-8">
 	<meta name="viewport"
 		content="width=device-width, initial-scale=1, minimum-scale=1, maximum-scale=1, user-scalable=no">
-	<title>XGo Builder</title>
+	<title>spx miniprogram player</title>
 	<style>
 		* {
 			margin: 0;
