+++ internal/webhost/static/runner.html
@@ -5,7 +5,7 @@
 	<meta charset="utf-8">
 	<meta name="viewport"
 		content="width=device-width, initial-scale=1, minimum-scale=1, maximum-scale=1, user-scalable=no">
-	<title>XGo Builder</title>
+	<title>spx web runtime</title>
 	<style>
 		html,
 		body {
@@ -54,8 +54,6 @@
 		let lastAssetURLs = null
 		let lastConfig = null
 
-		window.spxIsProxyThreadMode = false
-
 		function onProgress(value) {
 			const progressEvent = new CustomEvent('onProgress', {
 				detail: {
@@ -83,8 +81,8 @@
 		}
 
 		async function handleGodotExit(code) {
-            await restartEngineAfterCrash("godot exit", code);
-        }
+			await restartEngineAfterCrash("godot exit", code);
+		}
 
 		let restartAttempts = 0;
 		let onGameError = null
@@ -105,7 +103,7 @@
 			gameApp = null;
 			try {
 				await window.initEngine(lastAssetURLs, lastConfig);
-				restartAttempts = 0; 
+				restartAttempts = 0;
 				console.log("Engine reinitialized.");
 			} catch (e) {
 				console.error("Engine re-init failed:", e);
@@ -148,27 +146,17 @@
 
 		window.initEngine = async (assetURLs = null, config = {}) => {
 			isShowEditor = false
-			
+
 			if (gameApp != null) {
 				await gameApp.ResetGame();
 				return;
 			}
 
-			if(EnginePackMode == "worker"){
-				window.xbuilder_set_ai_description = function (description) {
-					gameApp.callWorkerFunction("xbuilder_set_ai_description", description)
-				}
-				window.xbuilder_set_ai_interaction_api_endpoint = function (endpoint) {
-					gameApp.callWorkerFunction("xbuilder_set_ai_interaction_api_endpoint", endpoint)
-				}
-				window.xbuilder_set_ai_interaction_api_token_provider = function (tokenProviderFunction) {
-					gameApp.callWorkerFunction("xbuilder_set_ai_interaction_api_token_provider", tokenProviderFunction)
-				}
-
-				window.on_wasm_loaded = function () {}
-				window.on_game_started = function () {}
+			if (EnginePackMode == "worker") {
+				window.on_wasm_loaded = function () { }
+				window.on_game_started = function () { }
 			}
-			
+
 			const finalConfig = {
 				'projectName': "spx_game",
 				'onProgress': onProgress,
@@ -185,7 +173,7 @@
 				"recordingOnGameStart": false,
 				"autoDownloadRecordedVideo": false,
 			};
-			
+
 			if (assetURLs != null) {
 				finalConfig.assetURLs = assetURLs
 			}
@@ -201,7 +189,7 @@
 			if (gameApp == null) throw new Error("gameApp is null, should call initEngine first")
 			await gameApp.InitGame(files)
 		}
-		
+
 		window.startGame = async () => {
 			if (gameApp == null) throw new Error("gameApp is null, should call initEngine first")
 			await gameApp.StartGame();
@@ -215,7 +203,7 @@
 			}
 			gameApp.startRecording()
 		}
-		
+
 		// Stops the current recording session and returns the result as a Blob.
 		window.stopRecording = async () => {
 			if (gameApp == null) {
@@ -291,7 +279,7 @@
 			}
 			gameApp.stepNextFrame();
 		}
-		
+
 		// Inform the parent window that the runner is ready (methods on the window object are callable)
 		window.dispatchEvent(new Event('runnerReady'));
 	</script>
